refactor: move useSignOut hook in AuthLayout folder
