+++ client/src/core/auth/hooks/index.ts
@@ -1,4 +1,3 @@
 export * from "./useAuth";
 export * from "./useRefreshToken";
 export * from "./useSecuredAxios";
-export * from "./useSignOut";
